removed errors

+++ StockExchangeSim/Code/Trading/Tools/LiquidityTool.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Eco
 {
@@ -19,18 +15,18 @@ namespace Eco
 
         public override LiquidityToolData StrategyOutcome(Company cp)
         {
-            
+
             if (cp.Dividend != MeasuredDividend)
             {
-                MeasuredDividend = (float)cp.Dividend;
+                MeasuredDividend = cp.Dividend;
                 PrevDividend = Dividend;
-                Dividend = (float)(cp.Dividend * (31556926.0 / Master.inst.SecondsPerTick));
+                Dividend = (cp.Dividend * (31556926.0f / Master.inst.SecondsPerTick));
             }
             if (PrevDividend == -1)
                 PrevDividend = Dividend;
 
             float MarketTrust = PrevDividend > 0 ? MathF.Pow(2, (Dividend - PrevDividend) / PrevDividend) : 0;
-            float StockPrice = 40 * Dividend * MathF.Pow((float)cp.CompetitivePosition, 3) * MarketTrust
+            float StockPrice = 40 * Dividend * MathF.Pow(cp.CompetitivePosition, 3) * MarketTrust
                 + MathF.Pow((float)cp.CompetitivePosition, 3) * (float)cp.Value * (float)cp.StockPart;
             return new LiquidityToolData(StockPrice);
         }
